Trait eclectic::map::VacantEntry

pub trait VacantEntry {
    type Key;
    type Value;
    type MutValue;
    fn insert(self: Box<Self>, value: Self::Value) -> Self::MutValue;
}

A vacant entry.

Associated Types

type Key

The type of the entry's key.

type Value

The type of the entry's value.

type MutValue

The type of the mutable reference to the entry's value with the same lifetime as the map.

Required Methods

fn insert(self: Box<Self>, value: Self::Value) -> Self::MutValue

Inserts the entry into the map with the given value.

Returns a mutable reference to the entry's value with the same lifetime as the map.